Add files via upload
This set of screenshots shows all of the basic and additional filter lists that are commonly used. The unchecked ones at the bottom are for specific sites such as Gelbooru & Kiss' suites of sites such as KissAnime, KimCartoon, Kissmanga, etc. Those are a must if you frequently use them as these sites have ads that can slow your browser down and even have trackers and viruses in them. The Nano Defender set are downloaded offsite and aren't part of the addon suite proper much like the Gelbolube & Kiss Filters.